Information and configuration for authorization
mu-authorization requires the setup of particular groups.  This is a
first draft to allow groups to be matched.  There are some notable
exceptions that bring us closer to making this work in our current
stack (development will continue soon).  Our configuartion currently
leaves some content in a shared application graph and does not clean
the application graph itself.