Shuffeled some things around to find a balance:
Previous version showed forms in status new or concept AND SMC without
forms.
The first part of the result, is not necessarily handy, as people
cannot make difference between broken and concept-state.

So worked with UNION, which will find the SMC without forms. And the
other part, made a little more strict as to find the forms users are
expecting.

The report will still show weird lines if something is broken IN the
form.
The report will still show lines if SMS without form.

But we will still miss some cases in the link between form and SMC,
e.g.
when form is not partOf a flow step.