Remove fields which are not relevant for the form.ttl
This form.json is an extra mapping file needed for the backend to
process the form.
No extra fields should be included here, as it has weird side effects.

In the current case, when tailored fields are extracted, there is
first a deletion of properties of the form that happens, and the
fields to be deleted are specified in form.json

The fields removed in this commit, were also part of the defintion and
thus removed. But these are required fields, and they basically broke
the form.

As a result, we're unable to send it out.

Hopefully this links stays relevant here for while

https://github.com/lblod/subsidy-applications-management-service/blob/master/lib/services/semantic-form-management.js#L153

I will also include a copy/paste of the code which deletes it, just to archive ite
```
       await new SourceDataExtractor().extract(form.uri,
          form.bundle.specification.definition);
        await this.mutator.delete(deprecated, form.graph);
```

Granted this very implict relations, should be improved in code. The
documentation has been updated accordingly in subsidy-management-service