Changed rangorde type to string, includes migration
Because of a bug in Virtuoso, strings with a language tags can, in some
cases, not be removed from the triplestore. To workaround this issue,
these particilar language tags can be removed with the help of a
migration, a change in the relevant resource configuration and some
changes in the frontend (PR on the way).