Tweak the conditions
This should properly display the correct field. The downside is that the field will only be shown once a value is selected for the executing authority level field.

It also requires us to list all the values that should show the "required" label which is easy to forget if we ever add new codelist values.