Use `http` instead of `https` for the rdfs prefix uri
It seems this is the version we use around the app and it was configured that way in the lpdc.json file. When we removed that file we missed that the prefix had a different protocol than the version in repository.lisp and it seemed that the lpdc.json version 'won' and we were implicitly depending on that in the app.