Bump clean-up-submission-service
This version includes a new API endpoint for removing a melding that is
still in concept status, and the existing removal API was rewritten
without mu-auth-sudo to improve security. For this, the mu-auth config
needed to be adapted. The dispatcher also has a new rule to allow for
the removal request as part of the Vendor API.