Get focused graph
The delta publication graph of leidinggevenden and mandatarissen are
still not in a separate publication-store.
So the previous migration would have implicilty changed these values
in the publication graph too. This would have broken the healing
mechanism (which makes a diff between source and 'what has been
published yet') and resulting in this change not being propagated to
- mandaten.lokaalbestuur.vlaanderen.be
- leidinggevenden.lokaalbestuur.vlaanderen.be

Anyway; it's a pain; we should move these two graphs to the separate
publicatoin triple store too