add form properties to first step form.json
The reason we do this is because the first step is the only step at which you can delete the form. When deleting the form, the form.json gets used to find the properties to delete. If you do not add these properties then they will be dangling in the database