Remove wrong labels for bestuurseenheden
Some bestuurseenheden had two labels: the correct one in the public graph, and a wrong one
in another (automatic submission, eredienstBedienaar and other graphs).

For the automatic submission graph, we delete the existing label and copy the correct one from
the public graph. For the other graphs, we simply delete the label there.