Migration: swap lang tag removal for re-sync
Instead of performing a removal of duplicates in triples where one has a
language tag, we opted to just start from a complete resync of the
harvester data. This is going to be more deploy work, but safer in terms
of data correctness, given that the previous errors in the consumer are
unpredictable.